# Support replicating multi-GPU modules