--- a/test/common_cuda.py
+++ b/test/common_cuda.py
@@ -7,6 +7,7 @@ from common_utils import TEST_WITH_ROCM, TEST_NUMBA
 
 TEST_CUDA = torch.cuda.is_available()
 TEST_MULTIGPU = TEST_CUDA and torch.cuda.device_count() >= 2
+TEST_GEQ4GPU = TEST_CUDA and torch.cuda.device_count() >= 4
 CUDA_DEVICE = TEST_CUDA and torch.device("cuda:0")
 # note: if ROCm is targeted, TEST_CUDNN is code for TEST_MIOPEN
 TEST_CUDNN = TEST_CUDA and (TEST_WITH_ROCM or torch.backends.cudnn.is_acceptable(torch.tensor(1., device=CUDA_DEVICE)))