@mrshenli 2019-04-02
> Do you want to add some more tests for _group_by_device specifically, or do you think it has enough coverage through the existing tests of replicate?

good point, I will add more tests.