@apaszke 2019-04-04
Why are we using lists for that, when you really want a device->device mapping, which is exactly the use case for dicts??