@apaszke 2019-04-04
I would really say that this is completely out of scope for our library. If you want multi-GPU modules you should probably just use DistributedDataParallel with a single process per module (i.e. no replication needed). I don't understand why are we happy to complicate the code this much for a use case which is already supported pretty well.