@mrshenli 2019-04-04
@apaszke 

> I would really say that this is completely out of scope for our library. If you want multi-GPU modules you should probably just use DistributedDataParallel with a single process per module (i.e. no replication needed). 

Some changes need to be added to DDP, otherwise it will try to replicate the input model using the given `device_ids`. There are several alternatives:

1. extends the current API for multi-GPU models: supports model replications in each process (this PR's approach).
2. have a limited API for multi-GPU models: tell users that if you are using multi-GPU models, we are going to ignore the `device_ids` arg, and don't do replications within individual processes at all.  (do you prefer this one?) 
3. disallow model replication in each DDP process for both single-GPU or multi-GPU models: this will help further simplify our DDP code.

BTW, why do we need model replications in a DDP process in the first place?