@mrshenli 2019-04-04
> Why are we using lists for that, when you really want a device->device mapping, which is exactly the use case for dicts??

I feel it's more natural to say I have two replicas, one on devices [0, 1], and another on devices [2, 3]. But I don't have a strong opinion on this one. If you like a map more, I can make the change.