@mrshenli 2019-04-04
> Did we benchmark the effects of this PR on nn.DataParallel? The code in torch/nn/parallel has already been a significant bottleneck for many models, and it got a lot more complex because of this PR!

Benchmark: use a very light model to maximize the impact of the new code. The code now is indeed more complex, but the execution path is roughly the safe for existing use cases. 

before:

```python
In [1]: import timeit

In [2]: import torch
   ...: from torch.nn.parallel import DataParallel as dp
   ...: 
   ...: def f():
   ...:     net = torch.nn.Linear(100, 100)
   ...:     net = dp(net.cuda(0))
   ...:     input = torch.rand(100, 100).cuda(0)
   ...:     for _ in range(20):
   ...:         net(input)
   ...: 

In [3]: %timeit f()
19 ms ± 488 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %timeit f()
19.4 ms ± 438 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [5]: %timeit f()
19.7 ms ± 373 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

after:

```python

In [3]: %timeit f()
19 ms ± 430 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %timeit f()
18.8 ms ± 113 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit f()
19.4 ms ± 957 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```