@apaszke 2019-04-07
Well the benchmark is not very useful, because you're not really measuring the forward times, but the time it takes to allocate weights, sample their values, initialize DP, etc. I just want to see the effect on the _iteration time_.

Regarding your points, I think we should have a limited API for multi-GPU modules. If there's no need to replicate the modules (i.e. for each instance of the model the user spins up a new process) then we should never do it! IIRC this is how it used to work when I wrote DDP, but maybe it has changed since then.